Implemented  ICS Export functionality
iCalExport class -
Created to handle export operations in an encapsulated manner
iCalExport takes an Event Object and will output an ICS File
Current export location is hardcoded and will deposit into the project
folder.
For later implementation this could be pointed towards a google calendar
API or similar export location.

MenuController.Java -
Added call for iCalExport, thus when calendar is created/selected the
system will export the ICS files automatically. Ideally, an enhancement
to this issue would be to create an "Export" button.
iExport.iCalSetup establishes the file folder for the export, this
stub/method could be used in API export implementation