Depress module intuitively based on mouse
If the mouse clicks on the module, depress it like normal. However, if the mouse leaves the area, the module goes back to normal even if the mouse is still depressed. If the mouse re-enters, the module goes back to pressed mode.
If the mouse enters normally, then show the hover effect. If the mouse leaves without being depressed, then return module shadows to normal.