Close notifications and menu when clicking on tables
When clicking on the tables in certain views, the hamburger menu and notification panel now close.

Also, settings the menu boolean to "open" when opening it fixes a bug where sometimes the application would 'forget' whether the menu was open.
In this commit, it's line 1443.

Finally, the notifications close when switching to another view.