Fix input validation on new user screen
This commit updates the input validation on the new user screen. Instead
of preventing the user from pressing okay and highlighting the input
errors, the user presses okay and an alert is displayed so the user can
better understand what they entered improperly. A confirmation dialog is
displayed, if the user leaves the name field empty. The name field is
optional, to allow for user privacy, so the user needs to confirm that
they don’t want to provide their name. The account number is now
represented by a student’s w number, to better link the program to
Wright State.
   Other minor fixes include:
 -Labeling the salutation and the name field as optional, for user
convenience.
-Giving an example w number in the account number, for user convenience.
Closes #223