Fixed output for NULL building and added serial ID
The previous output for NULL buildings would output a name and a
room number. A room number is generally not helpful information if
nobody knows what building it is in. This fixes Issue 249.

Also added a serialVersionUID based on the one found in Activity.java
so that Room.java uses the same serialization.