Remove secondary start-up Screen
This commit removes the screen that appears when the program first
starts. The new file and open file functionality is now in the main pane
of Raider Planner. This allows the user to open files and create new
files without exiting the program. It will also remove the odd startup
pane, that looks strange in some formats. This will allow the program to
run in a more standardized way, that users are used to, and what they
would expect in today’s market
Other fixes include:
When the program is run, by default it opens the last file that was
saved. In my opinion changing the opening welcome screen until a user
selects a file was out of the scope of this project.

If there are no files in the program, except the sample, the new account
screen will pop up automatically the first time the program is run.
Again, making a default welcome screen was out of the scope of this
issue.
closes #134