Adds a test class for Apache Derby database
This class is designed to test that Apache Derby is configured correctly with the project. The main function of this test class is to instantiate an in memory database and performs certain DB operations (e.g. create table, insert values, remove table). If this test executes successfully then we are assured that the Derby database is running correctly in our Gradle project.