Fixed input validation issues on new requirements window
Fixed issue that allowed a user to enter incorrect information into the
text boxes, but the ok/submit button would still unlock. This caused
exceptions and for the gui to not react when the user pressed the
ok/submit button. Also added code to verify that the quantity entry is
an integer.

Fixes #313