Display StudyProfile details in the Study Profiles window
Prior to this commit StudyProfile details were displayed in
a seperate pop-up window, this commit moves that information
to the same window from which it was called. This is more intuitive
for the user because it keeps all the information in one window.
The pop-up window still displays if you access the StudyProfile
through the notifications system.

Resolves: #362