Added ToolTips and improved data validation for New Activities
Added ToolTips to the ActivityController page to give feedback to users
about invalid input. Also improved validation for user input. Stopped
errors being thrown when a double was entered in the text fields.

Added isInteger() to MainController. This helps catch errors thrown when
a double (string) is being converted to an Integer.

Modified MilestoneController, RequirementController, and TaskController
to validate information with isInteger().

Fix: #376