No longer quits if user chooses not to create a new account-Fixes #406
Added boolean as argument for createAccount that is true if there is no
existing account. Changed method to return null if no new account is
created and an account already exists.
Added logic statements that abort the rest of the account creation
process if the user doesn't make a new account.
Changed reloadMain to this.main to prevent notification window from
opening if user cancels account creation.